Step 0: lib.rs split, OutputSink trait, test harness
Splits the crate into lib.rs (DagShell, OutputSink) and a thin main.rs
(binary entry point). Routes all DagShell command output through an
OutputSink so tests can capture it without touching stdout.

Adds tests/shell.rs with CapturingSink and the first integration test
(execute_routes_output_through_sink). No behaviour change.
